Editor: Move render options from sidebar to menubar.
