+++ editor/js/Sidebar.Project.js
@@ -3,8 +3,6 @@ import { UISpan } from './libs/ui.js';
 import { SidebarProjectApp } from './Sidebar.Project.App.js';
 /* import { SidebarProjectMaterials } from './Sidebar.Project.Materials.js'; */
 import { SidebarProjectRenderer } from './Sidebar.Project.Renderer.js';
-import { SidebarProjectImage } from './Sidebar.Project.Image.js';
-import { SidebarProjectVideo } from './Sidebar.Project.Video.js';
 
 function SidebarProject( editor ) {
 
@@ -15,14 +13,6 @@ function SidebarProject( editor ) {
 	/* container.add( new SidebarProjectMaterials( editor ) ); */
 
 	container.add( new SidebarProjectApp( editor ) );
-
-	container.add( new SidebarProjectImage( editor ) );
-
-	if ( 'SharedArrayBuffer' in window ) {
-
-		container.add( new SidebarProjectVideo( editor ) );
-
-	}
 
 	return container;
